Worker development
* Bring worker on par with web configuration
* Common Config TODO
