+++ app.js
@@ -236,53 +236,3 @@ async.waterfall([
 });
 
 module.exports = app;
-var mongoClient = require('mongodb');
-var path = require('path');
-var underscore = require('underscore');
-
-var handleMain = require(path.join(__dirname, 'handlers', 'main'));
-var handleMobile = require(path.join(__dirname, 'handlers', 'mobile'));
-var handleShare = require(path.join(__dirname, 'handlers', 'share'));
-
-var amqpURI = process.env.AMQP_URI || 'amqp://localhost';
-var mongoURI = process.env.MONGODB_URI || 'mongodb://localhost/VITacademics';
-
-var app = {
-    db: null,
-    queue: null
-};
-
-var onConnect = function (err, db) {
-    app.db = db;
-};
-mongoClient.connect(mongoURI, onConnect);
-
-var queue = jackrabbit(amqpURI);
-queue.queues = {
-  main: 'Main',
-  mobile: 'Mobile',
-  share: 'Share'
-};
-queue.on('connected', function () {
-  app.queue = queue;
-  var forEachQueue = function (elt, i, arr) {
-      var onReady = function () {
-        switch (elt) {
-          case queue.queues.main:
-            handleMain(app);
-            break;
-          case queue.queues.mobile:
-            handleMobile(app);
-            break;
-          case queue.queues.share:
-            handleShare(app);
-            break;
-          default:
-            console.log('Unsupported Queue: ' + elt);
-            break;
-        }
-      };
-      queue.create(elt, {prefetch: 500}, onReady);
-  };
-  underscore.values(queue.queues).forEach(forEachQueue);
-});
