+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "VITacademics",
-  "version": "0.98.50",
+  "version": "0.98.60",
   "description": "VITacademics API Server and Web App",
   "homepage": "https://vitacademics-dev.herokuapp.com/",
   "bugs": {
