+++ test/api-legacy-test.js
@@ -1,4 +1,4 @@
-#!/usr/bin/env node
+#!/usr/bin/env iojs
 
 /*
  *  VITacademics
